requirements.txt
Flask
Flask-CORS
Pillow
ultralytics
